BUGFIX: Changed logic to get start & end times for events in ICS feed.
I added a simple method specific to the current event model which
returns a tuple of the start & end time. The end time can be None if
duration wasn't set however.  This makes things a bit more readable
and handles NULL values in event.duration cleanly.

The downside is that events without duration show up as an hour long
(at least on Google Calendar) when imported since they do have a
time set.  This is as opposed to an all-day event, which might get
the user to actually read the event's details (which would hopefully
be more accurate that the hour long guess).